requirements.txt
requests~=2.28.1
python-dotenv~=1.0.0
pytest~=7.1.2
pre-commit~=2.20.0
